# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Vdoc API
  version: 0.1.0
  description: Machine-readable contract for the Vdoc v0.1 HTTP API. Projects own typed Documents, including OpenAPI and Markdown. REST handlers return HTTP 200 with the semantic result in the Vdoc envelope body.
servers:
  - url: /
security: []
...
paths:
...
  /api/v1/open/document-shares/{share_id}:
    parameters:
      - {$ref: '#/components/parameters/ShareID'}
    get:
      tags: [Document Shares]
      summary: Read minimal metadata for an authorized public document share.
      operationId: getPublicDocumentShare
      security: [{DocumentShareAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/VdocEnvelopeResponse'}
  /api/v1/open/document-shares/{share_id}/unlock:
    parameters:
      - {$ref: '#/components/parameters/ShareID'}
    post:
      tags: [Document Shares]
      summary: Exchange a valid share password for a short-lived memory-only unlock proof.
      operationId: unlockPublicDocumentShare
      security: [{DocumentShareAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/VdocEnvelopeResponse'}
  /api/v1/open/document-shares/{share_id}/versions:
    parameters:
      - {$ref: '#/components/parameters/ShareID'}
    get:
      tags: [Document Shares]
      summary: List published versions allowed by an all-versions share.
      operationId: listPublicDocumentShareVersions
      security: [{DocumentShareAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/VdocListEnvelopeResponse'}
  /api/v1/open/document-shares/{share_id}/versions/{version_id}/content:
    parameters:
      - {$ref: '#/components/parameters/ShareID'}
      - {$ref: '#/components/parameters/VersionID'}
    get:
      tags: [Document Shares]
      summary: Read an allowed immutable published snapshot.
      operationId: getPublicDocumentShareContent
      security: [{DocumentShareAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/VdocEnvelopeResponse'}
  /api/v1/open/document-shares/{share_id}/versions/{version_id}/download:
    parameters:
      - {$ref: '#/components/parameters/ShareID'}
      - {$ref: '#/components/parameters/VersionID'}
    get:
      tags: [Document Shares]
      summary: Download an allowed immutable published source file through Vdoc authorization.
      operationId: downloadPublicDocumentShareVersion
      security: [{DocumentShareAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/PublicDocumentDownloadResponse'}
...
  /api/v1/private/projects/{project_id}/documents:
    parameters:
      - {$ref: '#/components/parameters/ProjectID'}
    get:
      tags: [Documents]
      summary: List project documents.
      operationId: listProjectDocuments
      security: [{JWTAuth: []}]
      parameters:
        - name: document_type
          in: query
          required: false
          schema: {type: integer, minimum: 1}
          description: Filter documents by type code (1 OpenAPI, 2 Markdown).
      responses:
        '200': {$ref: '#/components/responses/VdocListEnvelopeResponse'}
    post:
      tags: [Documents]
      summary: Create project document.
      description: Creates a typed Document. document_type 1 is OpenAPI and 2 is Markdown. relative_path is the stored project-local path identity.
      operationId: createProjectDocument
      security: [{JWTAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/VdocEnvelopeResponse'}
  /api/v1/private/projects/{project_id}/documents/{document_id}:
    parameters:
      - {$ref: '#/components/parameters/ProjectID'}
      - {$ref: '#/components/parameters/DocumentID'}
    get:
      tags: [Documents]
      summary: Get project document.
      operationId: getProjectDocument
      security: [{JWTAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/VdocEnvelopeResponse'}
    patch:
      tags: [Documents]
      summary: Update project document.
      description: Updates document metadata only. Lifecycle status cannot be changed here; use the archive action. Invalid lifecycle input is rejected before any metadata is committed.
      operationId: updateProjectDocument
      security: [{JWTAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/VdocEnvelopeResponse'}
...
  /api/v1/private/projects/{project_id}/documents/{document_id}/shares:
    parameters:
      - {$ref: '#/components/parameters/ProjectID'}
      - {$ref: '#/components/parameters/DocumentID'}
    get:
      tags: [Document Shares]
      summary: List retained public-share history for a document.
      operationId: listDocumentShares
      security: [{JWTAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/VdocListEnvelopeResponse'}
    post:
      tags: [Document Shares]
      summary: Create an immutable branch-bound public-share policy and return its capability once.
      description: Requires an active Project, Document, and Branch with at least one published version. Archived parents retain list and revoke access but block creation.
      operationId: createDocumentShare
      security: [{JWTAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/VdocEnvelopeResponse'}
  /api/v1/private/projects/{project_id}/documents/{document_id}/shares/{share_id}/reveal:
    parameters:
      - {$ref: '#/components/parameters/ProjectID'}
      - {$ref: '#/components/parameters/DocumentID'}
      - {$ref: '#/components/parameters/ShareID'}
    post:
      tags: [Document Shares]
      summary: Reveal an active share capability without returning its password.
      description: Requires active parent resources. Archived parents retain list and revoke access but block capability reveal.
      operationId: revealDocumentShare
      security: [{JWTAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/VdocEnvelopeResponse'}
  /api/v1/private/projects/{project_id}/documents/{document_id}/shares/{share_id}/revoke:
    parameters:
      - {$ref: '#/components/parameters/ProjectID'}
      - {$ref: '#/components/parameters/DocumentID'}
      - {$ref: '#/components/parameters/ShareID'}
    post:
      tags: [Document Shares]
      summary: Irreversibly revoke an active public document share.
      operationId: revokeDocumentShare
      security: [{JWTAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/VdocEnvelopeResponse'}
...
  /api/v1/private/projects/{project_id}/documents/{document_id}/versions:
    parameters:
      - {$ref: '#/components/parameters/ProjectID'}
      - {$ref: '#/components/parameters/DocumentID'}
    get:
      tags: [Versions]
      summary: List published document versions.
      operationId: listDocumentVersions
      security: [{JWTAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/BranchIDQuery'}
      responses:
        '200': {$ref: '#/components/responses/VdocListEnvelopeResponse'}
...
  /api/v1/private/projects/{project_id}/documents/{document_id}/diffs:
    parameters:
      - {$ref: '#/components/parameters/ProjectID'}
      - {$ref: '#/components/parameters/DocumentID'}
    post:
      tags: [Diffs]
      summary: Compare document versions.
      description: Creates or returns a comparison only when the Project and Document are active and the two version IDs differ. Stored Diff history remains readable after archive.
      operationId: createDocumentDiff
      security: [{JWTAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/VdocEnvelopeResponse'}
    get:
      tags: [Diffs]
      summary: List stored document diffs.
      description: Optional from_version_id and to_version_id query parameters select an exact version pair.
      operationId: listDocumentDiffs
      security: [{JWTAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/VdocListEnvelopeResponse'}
...
  /api/v1/private/audit-logs:
    get:
      tags: [Audit Logs]
      summary: List authorized audit history.
      description: Super administrators may query all projects. Project administrators must supply project_id and can query only projects they administer. Reader and writer roles are denied. Results are newest first and capped at 200.
      operationId: listAuditLogs
      security: [{JWTAuth: []}]
      parameters:
        - {name: project_id, in: query, required: false, schema: {type: string}}
        - {name: action, in: query, required: false, schema: {type: string}}
        - {name: resource_type, in: query, required: false, schema: {type: string}}
        - {name: resource_id, in: query, required: false, schema: {type: string}}
        - {name: limit, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 200, default: 100}}
      responses:
        '200': {$ref: '#/components/responses/VdocListEnvelopeResponse'}
...
components:
  securitySchemes:
    JWTAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Raw JWT string issued by login or register.
    MCPTokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Raw MCP token string issued by the private MCP token APIs.
    DocumentShareAuth:
      type: apiKey
      in: header
      name: Authorization
      description: VdocShare scheme carrying the fragment-derived document-share capability. Protected links also require X-Vdoc-Share-Unlock after password exchange.
  parameters:
    BranchIDQuery:
      name: branch_id
      in: query
      required: false
      schema: {type: string, format: uuid}
      description: Filter drafts or versions to one document branch.
    ProjectID:
      name: project_id
      in: path
      required: true
      schema: {type: string}
      description: Project ID.
    DocumentID:
      name: document_id
      in: path
      required: true
      schema: {type: string}
      description: Document ID.
...
    VersionID:
      name: version_id
      in: path
      required: true
      schema: {type: string}
      description: Published document version ID.
    ShareID:
      name: share_id
      in: path
      required: true
      schema:
        type: string
        pattern: '^[0-9a-f]{32}$'
      description: Public document share ID. The capability secret is never placed in this path.
...
  responses:
    VdocEnvelopeResponse:
      description: HTTP 200 response with Vdoc envelope.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/VdocEnvelope'}
    VdocListEnvelopeResponse:
      description: HTTP 200 response with Vdoc envelope and total count.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/VdocListEnvelope'}
...
    PublicDocumentDownloadResponse:
      description: Authorized immutable raw OpenAPI or Markdown bytes with an attachment Content-Disposition header.
      headers:
        Content-Disposition:
          schema: {type: string}
      content:
        application/json:
          schema: {type: string, format: binary}
        application/yaml:
          schema: {type: string, format: binary}
        text/markdown:
          schema: {type: string, format: binary}
  schemas:
    VdocEnvelope:
      type: object
      properties:
        code: {type: integer}
        status: {type: string}
        description: {type: string}
        message: {type: string}
        trace_id: {type: string}
        timestamp: {type: integer}
        detail: {}
    VdocListEnvelope:
      allOf:
        - {$ref: '#/components/schemas/VdocEnvelope'}
        - type: object
          properties:
            total: {type: integer}
...